requirements.txt
pandas
requests
streamlit
langchain
langchain-openai
scikit-learn
matplotlib
wordcloud
joblib
tensorflow
konlpy
seaborn